@@ -5,17 +5,23 @@
         "build": "vite build && vite build --ssr"
     },
     "devDependencies": {
-        "@inertiajs/vue3": "^1.0.0",
-        "@tailwindcss/forms": "^0.5.3",
-        "@vitejs/plugin-vue": "^4.0.0",
-        "@vue/server-renderer": "^3.2.31",
-        "autoprefixer": "^10.4.12",
-        "axios": "^1.1.2",
-        "laravel-vite-plugin": "^0.7.2",
-        "lodash": "^4.17.19",
-        "postcss": "^8.4.18",
-        "tailwindcss": "^3.2.1",
-        "vite": "^4.0.0",
-        "vue": "^3.2.41"
+        "@inertiajs/vue3": "^1.0.11",
+        "@tailwindcss/forms": "^0.5.6",
+        "@vitejs/plugin-vue": "^4.3.4",
+        "@vue/server-renderer": "^3.3.4",
+        "autoprefixer": "^10.4.15",
+        "axios": "^1.5.0",
+        "laravel-vite-plugin": "^0.8.0",
+        "lodash": "^4.17.21",
+        "postcss": "^8.4.29",
+        "tailwindcss": "^3.3.3",
+        "vite": "^4.4.9",
+        "vue": "^3.3.4"
+    },
+    "dependencies": {
+        "@fortawesome/fontawesome-svg-core": "^6.4.2",
+        "@fortawesome/free-solid-svg-icons": "^6.4.2",
+        "@fortawesome/vue-fontawesome": "^3.0.3",
+        "vue3-easy-data-table": "^1.5.47"
     }
 }
