@@ -1,4 +1,6 @@
 import _ from 'lodash';
+import axios from 'axios';
+
 window._ = _;
 
 /**
@@ -7,7 +9,6 @@ window._ = _;
  * CSRF token as a header based on the value of the "XSRF" token cookie.
  */
 
-import axios from 'axios';
 window.axios = axios;
 
 window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
